mod: update AuthService paramaters

+++ MemeIt/MemeIt/Core/IAuthService.cs
@@ -4,7 +4,7 @@ namespace MemeIt.Core;
 
 public interface IAuthService
 {
-    AuthData GetAuthData(int id, string role);
+    AuthData GetAuthData(long id, string role);
     string HashPassword(string password);
     bool VerifyPassword(string? actualPassword, string hashedPassword);
 }
